Add the rel=canonical implementation from the Theme Directory.
